requirements.txt
# Student Notes Sharing Platform - Dependencies
# Core Flask packages
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-SocketIO>=5.3.6
Flask-Cors>=4.0.0

# Database
SQLAlchemy>=2.0.36

# WebSocket support
python-socketio>=5.10.0
python-engineio>=4.8.1

# File handling and image processing
Pillow>=10.0.0
PyMuPDF>=1.23.0
Werkzeug>=3.0.1

# Production server
gunicorn>=21.0.0

# Utilities
python-dotenv>=1.0.0
